refactor(services): move providers to contexts
