+++ app/src/components/Home/HomeBody.tsx
@@ -1,6 +1,6 @@
 import loadable from "@loadable/component";
 import { createStyles, makeStyles } from "@material-ui/core";
-import { OpenedFileContextProvider } from "@ufh/react-services/src/providers/OpenedFileContextProvider";
+import { OpenedFileContextProvider } from "@ufh/react-services/src/contexts/OpenedFileContextProvider";
 import React, { memo } from "react";
 import { useContextSelector } from "use-context-selector";
 import HomeBodyAddFiles from "./HomeBodyAddFiles";
